+++ src/50expression.js
@@ -422,6 +422,7 @@ yy.UniOp = function (params) { return yy.extend(this, params); }
 yy.UniOp.prototype.toString = function() {
 	if(this.op == '-') return this.op+this.right.toString();
 	if(this.op == '+') return this.op+this.right.toString();
+	if(this.op == '#') return this.op+this.right.toString();
 	if(this.op == 'NOT') return this.op+'('+this.right.toString()+')';
 	else if(this.op == null) return '('+this.right.toString()+')';
 };
@@ -440,6 +441,14 @@ yy.UniOp.prototype.toJavaScript = function(context, tableid, defcols) {
 	if(this.op == '-') return "(-("+this.right.toJavaScript(context, tableid, defcols)+"))";
 	if(this.op == '+') return "("+this.right.toJavaScript(context, tableid, defcols)+")";
 	if(this.op == 'NOT') return '!('+this.right.toJavaScript(context, tableid, defcols)+')';
+	if(this.op == '#') {
+		if(this.right instanceof yy.Column) {
+			return "(alasql.databases[alasql.useid].objects[\'"+this.right.columnid+"\'])";
+		} else {
+			return "(alasql.databases[alasql.useid].objects["
+				+this.right.toJavaScript(context, tableid, defcols)+"])";
+		};
+	}
 	else if(this.op == null) return '('+this.right.toJavaScript(context, tableid, defcols)+')';
 };